+++ experiments/automatic_scan_segmentation.py
@@ -16,8 +16,8 @@ import sys
 from pathlib import Path
 import datetime
 
-experiment_title = "scan segmentation default and each training"
-experiment_title = "scan segmentation default and each training with weight"
+experiment_title = "Whole Scan Segmentation default and each training"
+experiment_title = "Whole Scan Segmentation default and each training with weight"
 method = "GLCMTFS"
 
 experiment_datetime = datetime.datetime.now()
@@ -62,12 +62,12 @@ mainapp.set_persistent_cols({
     "Experiment Datetime": experiment_datetime.isoformat(" ", "seconds"),
 })
 
-mainapp.set_parameter("Processing;Scan Segmentation;Segmentation Method", method)
+mainapp.set_parameter("Processing;Whole Scan Segmentation;Segmentation Method", method)
 mainapp.set_parameter("Processing;Intensity Normalization", False)
 # mainapp.set_parameter("Processing;Intensity Normalization;Run Intensity Normalization", True)
-# mainapp.set_parameter("Processing;Scan Segmentation;Working Resolution", 0.000005) # default is 0.000001 (10 um)
-# mainapp.set_parameter("Processing;Scan Segmentation;Working Resolution", 0.0000025) # default is 0.000001 (10 um)
-mainapp.set_parameter("Processing;Scan Segmentation;Save Training Labels", True)
+# mainapp.set_parameter("Processing;Whole Scan Segmentation;Working Resolution", 0.000005) # default is 0.000001 (10 um)
+# mainapp.set_parameter("Processing;Whole Scan Segmentation;Working Resolution", 0.0000025) # default is 0.000001 (10 um)
+mainapp.set_parameter("Processing;Whole Scan Segmentation;Save Training Labels", True)
 # mainapp.set_parameter("Processing;Lobulus Segmentation;Central Vein Segmentation;Threshold", 0.18)
 mainapp.set_parameter("Processing;Lobulus Segmentation;Central Vein Segmentation;Threshold", 0.20)
 # mainapp.parameters.param("Processing", "Lobulus Segmentation", "Central Vein Segmentation", "Threshold").setValue(0.20)
@@ -93,16 +93,16 @@ def set_same(mainapp, fn):
     mainapp.set_parameter("Processing;Skeleton Analysis", False)
     mainapp.set_parameter("Processing;Texture Analysis", False)
     mainapp.set_parameter("Processing;Open output dir", False)
-    # mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Load Default Classifier", True)
-    mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Clean Before Training", False)
-    mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Training Weight", 100)
-    mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Run Training", True)
-    # mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Run Training", False)
-    mainapp.set_parameter("Processing;Scan Segmentation;Lobulus Number", 0)
-    mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Load Default Classifier", True)
-    mainapp.set_parameter("Processing;Scan Segmentation;Save Training Labels", True)
-    # mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Clean Before Training", False)
-    # mainapp.set_parameter("Processing;Scan Segmentation;TFS General;Training Weight", 100)
+    # mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Load Default Classifier", True)
+    mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Clean Before Training", False)
+    mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Training Weight", 100)
+    mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Run Training", True)
+    # mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Run Training", False)
+    mainapp.set_parameter("Processing;Whole Scan Segmentation;Lobulus Number", 0)
+    mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Load Default Classifier", True)
+    mainapp.set_parameter("Processing;Whole Scan Segmentation;Save Training Labels", True)
+    # mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Clean Before Training", False)
+    # mainapp.set_parameter("Processing;Whole Scan Segmentation;TFS General;Training Weight", 100)
     # mainapp.set_parameter("Processing;Lobulus Segmentation;Manual Segmentation", True)
 
 fns = [